Update Docs to include frontend development details. Update code to actually align with docs.
