+++ kolibri/core/assets/src/kolibri_core_app.js
