+++ kolibri/core/assets/src/mediator/mediator.js
