+++ kolibri/core/bundles.ini
@@ -1,2 +1,2 @@
 [kolibri_core]
-entry_file = assets/js/kolibri_core_app.js
+entry_file = assets/src/kolibri_core_app.js
